Moving extractSecuritySchemes into Common
